+++ pytorch_segmentation_models_trainer/tools/gee/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""GEE (Google Earth Engine) download tools for LULC products."""